requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Pillow==10.1.0
tensorflow==2.15.0
numpy==1.24.3
scikit-learn==1.3.2
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
redis==5.0.1
psycopg2-binary==2.9.9